fixes & added worldguard support

+++ src/main/java/com/ordwen/simpleblockfreeze/events/PlayerInteractListener.java
@@ -24,13 +24,10 @@ public class PlayerInteractListener implements Listener {
 
     @EventHandler
     public void onPlayerInteract(PlayerInteractEvent event) {
-        System.out.println("PlayerInteractListener.onPlayerInteract");
-
         final ItemStack item = event.getItem();
         if (item == null) return;
         if (!item.isSimilar(Configuration.getItem())) return;
 
-        System.out.println("Item is correct");
         event.setCancelled(true);
 
         final Player player = event.getPlayer();
@@ -43,21 +40,20 @@ public class PlayerInteractListener implements Listener {
         final Block block = event.getClickedBlock();
         if (block == null) return;
 
-        System.out.println("Block is not null");
-
         final Location location = block.getLocation();
         final World world = location.getWorld();
         if (world == null) return;
 
-        System.out.println("World is not null");
+        if (!Configuration.canBuild(player, world, location)) {
+            final String msg = Messages.UNAUTHORIZED_REGION.toString();
+            if (msg != null) player.sendMessage(msg);
+            return;
+        }
 
         final Action action = event.getAction();
-        System.out.println("Action: " + action);
         if (action == Action.RIGHT_CLICK_BLOCK) {
-            System.out.println("RIGHT_CLICK_BLOCK");
             sqlManager.deleteLocation(player, world.getName(), location.getX(), location.getY(), location.getZ());
         } else if (action == Action.LEFT_CLICK_BLOCK) {
-            System.out.println("LEFT_CLICK_BLOCK");
             sqlManager.saveLocation(player, world.getName(), location.getX(), location.getY(), location.getZ());
         }
     }
